# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FacturacionDemoAGVS API",
    "description": "API para la gestión de documentos y facturación.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local de desarrollo"
    }
  ],
  "paths": {
...
    "/empresa/{empresa_id}": {
      "get": {
        "tags": [
          "empresa"
        ],
        "summary": "Obtener una empresa por ID",
        "description": "Devuelve los detalles de una empresa específica identificada por su ID.",
        "parameters": [
          {
            "name": "empresa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la empresa obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "nombre": {
                      "type": "string",
                      "example": "Empresa XYZ"
                    },
                    "rif": {
                      "type": "string",
                      "example": "J-12345678-9"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Empresa no encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "empresa"
        ],
        "summary": "Actualizar una empresa por ID",
        "description": "Actualiza los detalles de una empresa específica identificada por su ID.",
        "parameters": [
          {
            "name": "empresa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Empresa Actualizada"
                  },
                  "rif": {
                    "type": "string",
                    "example": "J-98765432-1"
                  },
                  "domicilio_fiscal": {
                    "type": "string",
                    "example": "Av. Secundaria, Edificio ABC, Caracas"
                  },
                  "telefono": {
                    "type": "string",
                    "example": "+584121234568"
                  },
                  "email": {
                    "type": "string",
                    "example": "actualizado@empresa.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empresa actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "nombre": {
                      "type": "string",
                      "example": "Empresa Actualizada"
                    },
                    "rif": {
                      "type": "string",
                      "example": "J-98765432-1"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Empresa no encontrada"
          },
          "400": {
            "description": "Datos inválidos proporcionados"
          }
        }
      }
    },
...
    "/cliente/{cliente_id}": {
      "get": {
        "tags": [
          "cliente"
        ],
        "summary": "Obtener un cliente por ID",
        "description": "Devuelve los detalles de un cliente específico identificado por su ID.",
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del cliente obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "nombre": {
                      "type": "string",
                      "example": "Cliente XYZ"
                    },
                    "rif": {
                      "type": "string",
                      "example": "J-12345678-9"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "cliente"
        ],
        "summary": "Actualizar un cliente por ID",
        "description": "Actualiza los detalles de un cliente específico identificado por su ID.",
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Cliente Actualizado"
                  },
                  "rif": {
                    "type": "string",
                    "example": "J-98765432-1"
                  },
                  "domicilio_fiscal": {
                    "type": "string",
                    "example": "Av. Secundaria, Edificio ABC, Caracas"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "nombre": {
                      "type": "string",
                      "example": "Cliente Actualizado"
                    },
                    "rif": {
                      "type": "string",
                      "example": "J-98765432-1"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado"
          },
          "400": {
            "description": "Datos inválidos proporcionados"
          }
        }
      }
    },
...
    "/documento/{documento_id}": {
      "get": {
        "tags": [
          "documento"
        ],
        "summary": "Obtener un documento por ID",
        "description": "Devuelve los detalles de un documento específico identificado por su ID.",
        "parameters": [
          {
            "name": "documento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del documento obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "numero_control": {
                      "type": "string",
                      "example": "NC123456"
                    },
                    "empresa_id": {
                      "type": "integer",
                      "example": 101
                    },
                    "cliente_id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Documento no encontrado"
          }
        }
      }
    },
    "/documento/numero-control/{numero_control}": {
      "get": {
        "tags": [
          "documento"
        ],
        "summary": "Obtener un documento por número de control",
        "description": "Devuelve los detalles de un documento específico identificado por su número de control.",
        "parameters": [
          {
            "name": "numero_control",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NC123456"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del documento"
          },
          "404": {
            "description": "Documento no encontrado"
          }
        }
      }
    },
    "/documento/empresa/{empresa_id}": {
      "get": {
        "tags": [
          "documento"
        ],
        "summary": "Obtener documentos por ID de empresa",
        "description": "Devuelve una lista de documentos asociados a una empresa específica identificada por su ID.",
        "parameters": [
          {
            "name": "empresa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 101
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de documentos"
          },
          "404": {
            "description": "No se encontraron documentos para la empresa especificada"
          }
        }
      }
    },
    "/documento/cliente/{cliente_id}": {
      "get": {
        "tags": [
          "documento"
        ],
        "summary": "Obtener documentos por ID de cliente",
        "description": "Devuelve una lista de documentos asociados a un cliente específico identificado por su ID.",
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de documentos"
          },
          "404": {
            "description": "No se encontraron documentos para el cliente especificado"
          }
        }
      }
    },
    "/documento/create/factura": {
      "post": {
        "tags": [
          "documento"
        ],
        "summary": "Crear o recuperar una factura",
        "description": "Crea una nueva factura basada en un pedido existente identificado por su ID.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pedido_id": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": [
                  "pedido_id"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Factura creada o recuperada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "pedido_id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}